% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_SPARK_data.R
\name{add_bad_column_hosp}
\alias{add_bad_column_hosp}
\title{add_bad_column_hosp}
\usage{
add_bad_column_hosp(df, N, prob.of.bad.hosp)
}
\arguments{
\item{df}{argument}

\item{N}{argument}

\item{prob.of.bad.hosp}{argument}
}
\description{
add_bad_column_hosp
}
