% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_SPARK_data.R
\name{add_resistance}
\alias{add_resistance}
\title{add_resistance}
\usage{
add_resistance(df, N, ac.prob.list)
}
\arguments{
\item{df}{argument}

\item{N}{argument}

\item{ac.prob.list}{argument}
}
\description{
add_resistance
}
