% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_results.R
\name{delete_results}
\alias{delete_results}
\title{delete_results}
\usage{
delete_results(SpARKjags_model)
}
\arguments{
\item{SpARKjags_model}{a \code{string} specifying the model results to be
deleted or the directory in which all model results should be deleted}
}
\value{
Returns TRUE for each file deleted
}
\description{
Delete SpARKjags model results from within the SpARKjags package
}
\examples{
\dontrun{
# Delete single model results
delete_results("individual_models/test.rds")

# Delete all model results in a particular directory
delete_results("individual_models")
}

}
