% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICtable.R
\name{DICtable}
\alias{DICtable}
\title{DICtable}
\usage{
DICtable(models)
}
\arguments{
\item{models}{a \code{vector} of \code{strings}, each specifying the name of
a \code{runjags} object}
}
\value{
Returns a \code{flextable} containing DIC values for each model
output
}
\description{
DICtable
}
