% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathogen.R
\name{filter_pathogen}
\alias{filter_pathogen}
\title{filter_pathogen}
\usage{
filter_pathogen(data, pathogen)
}
\arguments{
\item{data}{data}

\item{pathogen}{pathogen}
}
\description{
filter_pathogen
}
