% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_violin.R
\docType{data}
\name{GeomSplitViolin}
\alias{GeomSplitViolin}
\title{https://stackoverflow.com/questions/35717353/split-violin-plot-with-ggplot2}
\format{
An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSplitViolin
}
\description{
https://stackoverflow.com/questions/35717353/split-violin-plot-with-ggplot2
}
\keyword{datasets}
