% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_human.R
\name{get_human}
\alias{get_human}
\title{get_human}
\usage{
get_human(
  classification,
  pathogen,
  indeterminate = "I",
  removeCarbapenem = FALSE,
  onlyCarbapenem = FALSE,
  removeQuinPen = TRUE
)
}
\arguments{
\item{classification}{Antibiotic classification}

\item{pathogen}{Pathogen name (Kleborate)}

\item{indeterminate}{c("R", "S"), default is "I"}

\item{removeCarbapenem}{default is FALSE; remove Carbapenem resitant samples}

\item{onlyCarbapenem}{default is FALSE}

\item{removeQuinPen}{default is TRUE}
}
\description{
get_human
}
