% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(model, data)
}
\arguments{
\item{model}{model}

\item{data}{data}
}
\value{
Returns a \code{data.frame} listing SpARK samples, and their
resistances to each antibiotic class. If model comes from the goodbad
directory, the posterior probability of being in the bad group (mean.p.bad)
is also included, which defines the badgroup (1 if mean.p.bad > 0.5)
}
\description{
import_data
}
