% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{list_models}
\usage{
list_models()
}
\value{
Returns a \code{data.frame} listing the model directories in the
SpARKjags package as well the models scripts and results contained within
them
}
\description{
List all models contained within model directories
}
\examples{
list_models()

}
