% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_autocorr.R
\name{plot_autocorr}
\alias{plot_autocorr}
\title{plot_autocorr}
\usage{
plot_autocorr(model, var.regex, filename)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{var.regex}{a regex \code{string} to filter variables}

\item{filename}{a \code{string} specifying the filename}
}
\description{
plot_autocorr
}
\examples{
\dontrun{
res.a_naive <- get_model("a_naive", "goodbad_models")
plot_autocorr(model = res.a_naive,
              var.regex = "(a.prob)|(intercept)|(sd)")
}

}
