% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caterpillar.R
\name{plot_caterpillar}
\alias{plot_caterpillar}
\title{plot_caterpillar}
\usage{
plot_caterpillar(model, save_to, model_name, save_pdf = FALSE)
}
\arguments{
\item{model}{model}

\item{save_to}{dir}

\item{model_name}{model name}

\item{save_pdf}{boolean}
}
\description{
plot_caterpillar
}
