% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{plot_correlation}
\usage{
plot_correlation(model, data)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{data}{a \code{list} containing the data input for the runjags model,
generated with \code{jags_data()}}
}
\description{
plot_correlation
}
