% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jags.R
\name{plot_jags}
\alias{plot_jags}
\title{plot_jags}
\usage{
plot_jags(res, var, labels, intercept, condense = F)
}
\arguments{
\item{res}{res}

\item{var}{var}

\item{labels}{labels}

\item{intercept}{intercept}

\item{condense}{condense}
}
\description{
plot_jags
}
