% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAC.R
\name{plotAC}
\alias{plotAC}
\title{plotAC}
\usage{
plotAC(model, var.regex, filename)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{var.regex}{a regex \code{string} to filter variables}

\item{filename}{a \code{string} specifying the filename}
}
\description{
plotAC
}
\examples{
res.a_naive <- get_model("a_naive", "goodbad_models")
plotAC(model = res.a_naive,
       var.regex = "(a.prob)|(intercept)|(sd)")

}
