% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_custom_model.R
\name{run_custom_model}
\alias{run_custom_model}
\title{run_custom_model}
\usage{
run_custom_model(data, custom_model, thin = 1, save_to)
}
\arguments{
\item{data}{a \code{list} output from the jags_data function}

\item{custom_model}{a \code{string} specifying a model location}

\item{thin}{(optional; default = 1) an \code{integer} specifying (from
\code{runjags::run.jags()}) the thinning interval to be used in JAGS.
Increasing the thinning interval may reduce autocorrelation, and therefore
reduce the number of samples required, but will increase the time required to
run the simulation. Using this option thinning is performed directly in JAGS,
rather than on an existing MCMC object.}

\item{save_to}{(optional) a \code{string} specifying a custom save location
for the model output. If save_to is missing, the model output will be
saved in the same directory as the model script.}
}
\value{
Returns code{string} specifying the location of the model output
}
\description{
Run custom model. If model results already exists, return file location.
}
\examples{
\dontrun{
run_model(data, "individual_models/a.R")
}

}
