% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(data, SpARKjags_model, thin, user_model, save_to)
}
\arguments{
\item{data}{data}

\item{SpARKjags_model}{model path e.g. goodbad_models/a.R}

\item{thin}{thin}

\item{user_model}{user_model}

\item{save_to}{save_to}
}
\description{
run_model
}
