% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_samples.R
\name{summarise_samples}
\alias{summarise_samples}
\title{summarise_samples}
\usage{
summarise_samples(model, data)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{data}{data}
}
\description{
Counts the number of samples identified as being in the good group or the
bad group, extracted from posterior probabilities
}
