% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPSRF.R
\name{testPSRF}
\alias{testPSRF}
\title{Potential Scale Reduction Factor}
\usage{
testPSRF(model, plot = T)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{plot}{a \code{boolean} object specifying whether or not a plot should
be generated}
}
\description{
The Gelman and Rubin (1992) Potential Scale Reduction Factor (PSRF, or
\eqn{\hat{R}}) is used to assess MCMC convergence. If the chains have not
converged to a common distribution, the \eqn{\hat{R}} statistic will be
greater than one. Values close to 1 are considered good.
}
