% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{traceplot}
\usage{
traceplot(model, var.regex, filename)
}
\arguments{
\item{model}{model}

\item{var.regex}{var.regex}

\item{filename}{filename}
}
\description{
traceplot
}
