% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wards}
\alias{wards}
\title{wards}
\format{
An object of class \code{data.frame} with 138 rows and 9 columns.
}
\usage{
wards
}
\description{
A dataset
}
\examples{
data("wards", package = "SpARKjags")

}
\keyword{datasets}
