% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibioticsplot.R
\name{antibioticsplot}
\alias{antibioticsplot}
\title{antibioticsplot}
\usage{
antibioticsplot(model, data)
}
\arguments{
\item{model}{model}

\item{data}{data}
}
\description{
Identify bad group samples from bad.p and look at location, amr profile, and
general distribution
}
