% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antibiotics.R
\name{plot_antibiotics}
\alias{plot_antibiotics}
\title{plot_antibiotics}
\usage{
plot_antibiotics(model, data)
}
\arguments{
\item{model}{model}

\item{data}{data}
}
\description{
Identify bad group samples from bad.p and look at location, amr profile, and
general distribution
}
