% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_autocorr.R
\name{plot_autocorr}
\alias{plot_autocorr}
\title{plot_autocorr}
\usage{
plot_autocorr(model, filename)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{filename}{a \code{string} specifying the filename}
}
\description{
plot_autocorr
}
