% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caterpillar.R
\name{plot_caterpillar}
\alias{plot_caterpillar}
\title{plot_caterpillar}
\usage{
plot_caterpillar(model, var.regex, filename)
}
\arguments{
\item{model}{model}

\item{var.regex}{var.regex}

\item{filename}{filename}
}
\description{
plot_caterpillar
}
