% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(model, data, var.regex)
}
\arguments{
\item{model}{a \code{runjags} object containing model results}

\item{data}{data input}

\item{var.regex}{a regex \code{string} to filter variables}
}
\description{
Generates a violin plot where violins represent the posterior probability of
resistance and points represent the calculated probability of resistance to
each antibiotic class.
}
\examples{
\dontrun{
data <- jags_data(classification = "all",
                  categories = "human",
                  pathogen = "Klebsiella pneumoniae",
                  removeQuinPen = TRUE)
res.a <- get_model("a", "goodbad_models")
plot_density(model = res.a,
            data = data,
            var.regex = get_vars(res.a))
}

}
