% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(data, model_name, directory, thin)
}
\arguments{
\item{data}{data}

\item{model_name}{model_name}

\item{directory}{directory}

\item{thin}{thin}
}
\description{
run_model
}
